requirements.txt
pandas
numpy
yfinance
gspread
google-auth
python-dotenv
scikit-learn
requests
schedule
openpyxl
